destination-aws-datalake: Add selector for JSON serialization library
Tables in AWS Glue require setting the classpath of the library that will be used by the
query engine to read and write the data being stored. The default is to use the OpenX
version, which was hardcoded in the connector. This adds a selector in the configuration
to select from the OpenX version used by AWS Athena, and the Hive version used by
engines such as Trino.