destination-aws-datalake: Add support for namespace prefix
It is desirable to be able to populate the S3 storage layer in a namespaced fashion to
reduce clutter at the top level of the destination bucket. This passes through the
`namespace` attribute of the configured stream to allow for specifying where the written
files are stored, letting the user organize their bucket layout how they see fit.