destination-aws-datalake: Add namespace support and aggressive purge
- Namespacing in S3 is equivalent to additional path prefixes in the key name. This
passes through the namespace attribute and incorporates it into the destination S3 path
to make it cleaner to track objects by which source they are generated from.
- The default behavior of LakeFormation is to soft delete objects, but that is
incompatible with how query engines other than Athena expect to interact with the
underlying S3 data. This adds a step to aggressively purge object data when operating in
`full refresh` mode.