destination-aws-datalake: Increase the session length for STS sessions
When using an IAM role for the aws-datalake destination connector it will default to a
session length of 1 hour. If the total length of time required to sync a connector to
that destination exceeds the 1 hour length the whole sync will fail due to an
`ExpiredTokenException`. This increases the session length to the maximum value of 12
hours to allow for long-running syncs.